requirements.txt
tensorflow-gpu==1.14
keras==2.2.4
dictor
pygame
tabulate
textx
sklearn
lxml
numba
natsort
pandas
ruamel_yaml
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.0.0
certifi==2018.8.24
chardet==3.0.4
click==7.1.1
configparser==4.0.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
elementpath==1.3.3
entrypoints==0.3
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
gql==0.2.0
graphql-core==1.1
grpcio==1.27.2
idna==2.9
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.2
notebook==6.0.3
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
open3d==0.9.0.0
opencv-python==4.2.0.32
pandocfilters==1.4.2
parso==0.6.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.8.1
pytorch-lightning==0.7.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
sentry-sdk==0.14.2
Shapely==1.7.0
shortuuid==1.0.1
six==1.14.0
smmap==3.0.1
subprocess32==3.5.4
tensorboard==1.14.0
terminado==0.8.3
testpath==0.4.4
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wandb==0.8.29
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
xmlschema==1.0.18
zipp==3.1.0
h5py==2.10.0
